chore: add storage layer with boltdb

+++ internals/node/genesis.go
@@ -1,6 +1,6 @@
 package node
 
 type Genesis struct {
-	Chain    string             `json:"chainName"`
-	Balances map[string]float32 `json:"balances"`
+	Chain    string            `json:"chainName"`
+	Balances map[string]uint32 `json:"balances"`
 }
